+++ RELEASE_NOTES.md
@@ -71,6 +71,8 @@ the Apache Avro binary format.
 
 * ADDED: New `base32`, `base58`, and `base85` libraries.
 
+* ADDED: New `json_schema` library.
+
 * ADDED: New `url` library with predicates for validating, parsing, generating,
 and normalizing URLs.